Enforce documentation of component attributes
