+++ app/cells/markdown_readonly/markdown_readonly.yml
@@ -1,0 +1,7 @@
+attributes:
+  content:
+    description: 'Markdown text.'
+
+examples:
+  - attributes:
+      content: "# Heading\n\nLorem ipsum..."
