+++ lib/ui_components.rb
@@ -3,6 +3,7 @@ Gem.loaded_specs['ui_components'].dependencies.each do |gem|
 end
 require 'active_model/railtie'
 require 'ui_components/engine'
+require 'ui_components/docu_cop'
 require 'ui_components/cell'
 require 'ui_components/form_helper'
 require 'ui_components/styleguide'
